% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_build_card_html.R
\name{build_card_html}
\alias{build_card_html}
\title{Build HTML Tags for Flash Card}
\usage{
build_card_html(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
data.frame
}
\description{
Build HTML Tags for Flash Card
}
\examples{
\dontrun{
 html <- build_card_html(card_data)
}
}
