% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_browse_deck.R
\name{mod_browse_deck_server}
\alias{mod_browse_deck_server}
\title{'Browse Deck' Module - Server}
\usage{
mod_browse_deck_server(id, topic)
}
\arguments{
\item{id}{character - module id}

\item{topic}{reactive data.frame - selected topic data}
}
\value{
reactive list
}
\description{
'Browse Deck' Module - Server
}
\examples{
\dontrun{
 selected_card <- mod_browse_deck_server('browse_deck')
}
}
