% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_browse_deck.R
\name{mod_browse_deck_ui}
\alias{mod_browse_deck_ui}
\title{'Browse Deck' Module - UI}
\usage{
mod_browse_deck_ui(id)
}
\arguments{
\item{id}{character - module id}
}
\value{
shiny tags
}
\description{
'Browse Deck' Module - UI
}
\examples{
\dontrun{
 mod_browse_deck_ui('browse_deck')
}
}
