% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_card_tbl.R
\name{mod_card_tbl_server}
\alias{mod_card_tbl_server}
\title{'Card Table' Module - Server}
\usage{
mod_card_tbl_server(id, card)
}
\arguments{
\item{id}{character - module id}

\item{card}{reactive data.frame - selected card data}
}
\description{
'Card Table' Module - Server
}
\examples{
\dontrun{
 mod_card_tbl_server('show_card', card)
}
}
