% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_card_tbl.R
\name{mod_card_tbl_ui}
\alias{mod_card_tbl_ui}
\title{'Card Table' Module - UI}
\usage{
mod_card_tbl_ui(id)
}
\arguments{
\item{id}{character - module id}
}
\value{
shiny tags
}
\description{
'Card Table' Module - UI
}
\examples{
\dontrun{
 mod_card_tbl_ui('show_card')
}
}
