% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_load_deck.R
\name{mod_load_deck_server}
\alias{mod_load_deck_server}
\title{'Load Deck' Module - Server}
\usage{
mod_load_deck_server(id)
}
\arguments{
\item{id}{character - module id}
}
\value{
reactive data.frame
}
\description{
'Load Deck' Module - Server
}
\examples{
\dontrun{
 load_deck <- mod_load_deck_server('load_deck')
}
}
