% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_load_deck.R
\name{mod_load_deck_ui}
\alias{mod_load_deck_ui}
\title{'Load Deck' Module - UI}
\usage{
mod_load_deck_ui(id)
}
\arguments{
\item{id}{character - module id}
}
\value{
shiny tags
}
\description{
'Load Deck' Module - UI
}
\examples{
\dontrun{
 module_load_deck_ui('load_deck')
}
}
