% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_show_card.R
\name{mod_show_card_server}
\alias{mod_show_card_server}
\title{'Show Card' Module - Server}
\usage{
mod_show_card_server(id, card)
}
\arguments{
\item{id}{character - module id}

\item{card}{reactive data.frame - selected card data}
}
\description{
'Show Card' Module - Server
}
\examples{
\dontrun{
 mod_show_card_server('show_card')
}
}
