% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_show_card.R
\name{mod_show_card_ui}
\alias{mod_show_card_ui}
\title{'Show Card' Module - UI}
\usage{
mod_show_card_ui(id)
}
\arguments{
\item{id}{character - module id}
}
\value{
shiny tags
}
\description{
'Show Card' Module - UI
}
\examples{
\dontrun{
 mod_show_card_ui('show_card')
}
}
